realtime/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
redis==5.0.3
PyJWT==2.8.0
pydantic==2.6.4
python-dotenv==1.0.1
asyncpg==0.29.0
pytest==8.1.1
pytest-asyncio==0.23.6
